Added the most complete documentation I could find. Also corrected a bug in the motor controller function executeAfterMS, where it would clear the other pins on the PORT. Added overflow functionality to executeAfterMS and tested with blinking LEDs. Note: was useful for testing, but might not be present on the final version, since it introduces extra time on the interrupts.

+++ timer.h
@@ -3,7 +3,7 @@
 
 #define T_CLK (1000000l)
 #define DIV (8l)
-#define TICKS_PER_MS (T_CLK/(DIV*1000))
+#define TICKS_PER_MS ((int)(T_CLK/(DIV*1000)))
 
 extern long int timerMS;
